+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The public API of this library consists of the functions declared in file
 [h3api.h.in](./src/h3lib/include/h3api.h.in).
 
 ## [Unreleased]
+### Added
+- `h3ToCenterChild` function to find center child at given resolution (#263)
+- `getPentagonIndexes` (and `pentagonIndexCount`) function to find pentagons at given resolution (#263)
 
 ## [3.5.0] - 2019-07-22
 ### Added